Time: 289 ms (18.28%), Space: 77.3 MB (91.50%) - LeetHub

+++ 1339-maximum-product-of-splitted-binary-tree/1339-maximum-product-of-splitted-binary-tree.cpp
@@ -1,9 +1,9 @@
 class Solution {
 public:
     long int ans = 0 , total = 0;
-    long int solve(TreeNode* root){
+    int solve(TreeNode* root){
         if(!root) return 0;
-        long int cur = solve(root->left)+solve(root->right)+root->val;
+        int cur = solve(root->left)+solve(root->right)+root->val;
         ans = max(ans,cur*(total-cur));
         return cur;
     }
